Move skills up page to group non-nav sections (#30)
To follow #28 as a temporary fix to the nav bar, I'm moving skills up the screen to be near the other sections that are not listed on the nav bar. This allows the sections with a jump button on the nav bar to be further down the screen and those without to be grouped at the to for visibility all around.